Upgrade: 토큰 유호성 인증 미들웨어 로직 보완 * 리프레시토큰이 존재하지 않을 때, 로그인할 수 있도록 json 메시지 전달

+++ utils/authenticateToken.ts
@@ -26,6 +26,15 @@ const authenticateToken = (req: Request, res: Response, next: NextFunction) => {
           UserUserId: decoded!.id
         }
       });
+
+      if (!refreshToken) {
+        return res
+          .status(401)
+          .json({
+            message: '로그인이 필요합니다.'
+          });
+      }
+
       jwt.verify(refreshToken!.dataValues.refreshToken, process.env.JWT_REFRESH_TOKEN!, async (error: any, user: any) => {
         if (error) {
           await RefreshToken.destroy({
